% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{delta_mu}
\alias{delta_mu}
\title{Conditional Average Treatment Effect Estimator for Y}
\usage{
delta_mu(X)
}
\arguments{
\item{X}{A matrix of covariates.}
}
\value{
A numeric vector representing the estimated treatment effect for Y for each observation.
}
\description{
Computes the difference in expected Y outcomes under treatment and control, using \code{h_Y}.
}
\examples{
X <- matrix(runif(10*2), 10, 2)
delta_mu(X)
}
