% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{delta_mu_linear}
\alias{delta_mu_linear}
\title{Linear-shaped Conditional Average Treatment Effect estimator for Y}
\usage{
delta_mu_linear(X)
}
\arguments{
\item{X}{A matrix of covariates of size n x d (input data in \link{0,1}).}
}
\value{
A numeric vector that represents the contrast between primary outcomes for given \code{X}.
}
\description{
Computes the difference in expected Y outcomes under treatment and control, using \code{h_Y}.
}
\examples{
X <- matrix(runif(10*2), 10, 2)
delta_mu(X)
}
