% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{delta_nu_complicated}
\alias{delta_nu_complicated}
\title{Complicated Conditional Average Treatment Effect Estimator for Xi}
\usage{
delta_nu_complicated(X)
}
\arguments{
\item{X}{A matrix of covariates.}
}
\value{
A numeric vector representing the estimated treatment effect for Xi for each observation.
}
\description{
Computes the difference in expected outcomes under treatment and control.
}
\examples{
X <- matrix(runif(10*2), 10, 2)
delta_nu_complicated(X)
}
