% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{delta_nu_threshold}
\alias{delta_nu_threshold}
\title{Thresholded Conditional Average Treatment Effect estimator for Xi}
\usage{
delta_nu_threshold(X)
}
\arguments{
\item{X}{A matrix of covariates of size n x d (input data).}
}
\value{
A numeric vector that represents the contrast between adverse event outcomes for given \code{X}.
}
\description{
Computes the difference in expected outcomes under treatment and control.
}
\examples{
X <- matrix(runif(10*2), 10, 2)
delta_nu(X)
}
