% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimate_ps}
\alias{estimate_ps}
\title{Estimate propensity score}
\usage{
estimate_ps(
  A,
  X,
  folds,
  SL.library = c("SL.glm", "SL.mean"),
  V = 2L,
  threshold = 0.01
)
}
\arguments{
\item{A}{A binary vector or matrix of length n indicating treatment assignment (0 or 1).}

\item{X}{A matrix or data frame of covariates of size n x d (input data in \verb{[0,1]}).}

\item{folds}{A list of cross-validation folds (e.g., a list of indices for each fold).}

\item{SL.library}{Vector of libraries for training SuperLearner (c("SL.glm", "SL.mean") by default).}

\item{V}{Number of folds inside the SuperLearner (2L by default).}

\item{threshold}{A numeric scalar that sets the minimum allowed value for upper and lower bound estimations (1e-2 by default). Constrains estimation to \verb{[threshold, 1 - threshold]}.}
}
\value{
A fold-specific function predicting propensity score given treatment (A) and covariates (X)
}
\description{
This function trains the propensity score models
using \code{SuperLearner}, applying cross-validation to compute out-of-fold estimates.
}
