% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimate_real_valued_mu}
\alias{estimate_real_valued_mu}
\title{Estimate real-valued mu}
\usage{
estimate_real_valued_mu(Y, A, X, folds, SL.library, V = 2L)
}
\arguments{
\item{Y}{A numeric vector or matrix of length n representing primary outcomes (in R).}

\item{A}{A binary vector or matrix of length n indicating treatment assignment (0 or 1).}

\item{X}{A matrix or data frame of covariates of size n x d (input data in R).}

\item{folds}{A list of cross-validation folds, typically created with \code{SuperLearner::CVFolds}.}

\item{SL.library}{Vector of libraries for training SuperLearner.}

\item{V}{Number of folds inside the SuperLearner (2L by default).}
}
\value{
A fold-specific function predicting primary outcome (Y) given treatment (A) and covariates (X)
}
\description{
This function trains conditional mean of primary outcome models for treated and control groups
using \code{SuperLearner}, applying cross-validation to compute out-of-fold estimates.
}
