% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expit}
\alias{expit}
\alias{logit}
\title{The expit/logit functions}
\usage{
expit(q, location = 0, scale = 1, lower.tail = TRUE, log.p = FALSE)

logit(p, location = 0, scale = 1, lower.tail = TRUE, log.p = FALSE)
}
\description{
expit(x) returns \code{1/(1+exp(-x))} for any real x.

logit(p) return \code{log(p/(1-p))} for any probability p.
}
