% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{generate_data}
\alias{generate_data}
\title{Synthetic data generator and functions generator}
\usage{
generate_data(
  n,
  ncov = 10L,
  scenario_mu = c("Linear", "Threshold", "Mix", "Null", "Constant"),
  scenario_nu = c("Linear", "Threshold", "Mix", "Satisfied"),
  is_RCT = FALSE,
  seed = NA
)
}
\arguments{
\item{n}{Number of observations to generate.}

\item{ncov}{Number of baseline covariates (at least 2L and 10L by default).}

\item{scenario_mu}{String indicating the type of scenario for delta_Mu ("Linear", "Threshold", "Mix", "Null", "Constant").}

\item{scenario_nu}{String indicating the type of scenario for delta_Nu ("Linear", "Threshold", "Mix", "Satisfied").}

\item{is_RCT}{Logical value indicating whether the scenario is an RCT (FALSE by default).}

\item{seed}{Integer or NA (NA by default).}
}
\value{
A list containing two data frames (\code{df_complete} with all potential outcomes and
treatment assignments and  \code{df_obs} with observed outcomes based on treatment) and the oracular
functions delta_Mu and delta_Nu.
}
\description{
Generates a dataset simulating treatment assignment, covariates, and potential outcomes.
}
\examples{
data <- data_gen(100)
head(data[[1]])  # complete data
head(data[[2]])  # observed data
}
