% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_results.R
\name{get_opt_beta_lambda}
\alias{get_opt_beta_lambda}
\title{Select Optimal Beta and Lambda Combination}
\usage{
get_opt_beta_lambda(combinations, root.path)
}
\arguments{
\item{combinations}{A matrix or data frame where each row corresponds to a beta and optimal-lambda pair.}

\item{root.path}{Path to the folder where all results are to be saved.}
}
\value{
A vector of intermediate results including the optimal: lambda, beta, risk, constraint, obj.
}
\description{
This function loads intermediate results corresponding to lambda-optimal solutions for each beta value.
It identifies and returns the beta-lambda combination that minimizes the objective function.
}
