% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_functions.R
\name{grad_Lagrangian_p}
\alias{grad_Lagrangian_p}
\alias{grad_Lagrangian_p_X}
\title{Gradient of the objective function}
\usage{
grad_Lagrangian_p(
  psi,
  X,
  delta_Mu,
  delta_Nu,
  lambda,
  alpha = 0.1,
  beta = 0.05,
  centered = FALSE
)

grad_Lagrangian_p_X(
  psi_X,
  delta_Mu_X,
  delta_Nu_X,
  lambda,
  alpha = 0.1,
  beta = 0.05,
  centered = FALSE
)
}
\arguments{
\item{psi}{A function that takes an input \code{X} and returns a numeric vector with values in the range \code{[-1, 1]}.}

\item{X}{A matrix of covariates of size n x d (input data in \link{0,1}).}

\item{delta_Mu}{A function of \code{X} that determines the contrast between primary outcomes.}

\item{delta_Nu}{A function of \code{X} that determines the contrast between adverse event outcomes.}

\item{lambda}{A non-negative numeric scalar controlling the penalty for violating the constraint.}

\item{beta}{A non-negative numeric scalar controlling the sharpness of the probability function (0.05 by default).}

\item{centered}{A logical value indicating whether to apply centering in \code{sigma_beta} (FALSE by default).}
}
\value{
A numeric vector representing the gradient of the objective function with respect to \code{psi(X)}.
}
\description{
Computes the gradient of the objective function with respect to \code{psi} at X.
The gradient is used in optimization algorithms like Stochastic Gradient Descent (SGD).
}
