% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{h_Y_complicated}
\alias{h_Y_complicated}
\title{h_Y_complicated: Complicated Treatment Effect on Y Component Function}
\usage{
h_Y_complicated(X, A)
}
\arguments{
\item{X}{A matrix of covariates where each row is an observation and columns represent features.}

\item{A}{A vector indicating treatment assignment (+1 or -1) for each observation.}
}
\value{
A numeric vector with the transformed values based on covariates and treatment.
}
\description{
Computes a linear interaction term between covariates and treatment.
}
\examples{
X <- matrix(runif(10*2), 10, 2)
A <- rep(1, 10)
h_Y_complicated(X, A)
}
