% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{iterative_psi_evolution}
\alias{iterative_psi_evolution}
\title{Visualize the evolution of the solutions psi obtained iteratively.}
\usage{
iterative_psi_evolution(
  intermediate_result,
  theta_opt,
  theta_t,
  X_test,
  beta = 0.05,
  centered = FALSE,
  root.path,
  name
)
}
\arguments{
\item{intermediate_result}{File from Intermediate folder gathering results from alternated procedure.}

\item{theta_opt}{The real numeric matrix (k x d). Each row is from oracular FW inner minimization, used to recover an extremal point for convex function construction.}

\item{theta_t}{An estimated numeric matrix (k x d). Each row is from T-learner FW inner minimization, used to recover an extremal point for convex function construction.}

\item{X_test}{A matrix of covariates from test data, of size n_test x d (input data).}

\item{beta}{A non-negative numeric scalar controlling the sharpness of the probability function (0.05 by default).}

\item{centered}{A logical value indicating whether to apply centering in \code{sigma_beta} (FALSE by default).}

\item{root.path}{Path to the folder where images are to be saved.}

\item{name}{A string to add to the end of filename.}
}
\value{
A message indicating that the image was saved.
}
\description{
Plot of the density of tje sigma_psi solutions obtained iteratively and the qq plot of the .
}
