% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_functions.R
\name{Lagrangian_p}
\alias{Lagrangian_p}
\title{Objective Function taking the form of a Lagrangian}
\usage{
Lagrangian_p(
  psi,
  X,
  delta_Mu,
  delta_Nu,
  lambda,
  alpha = 0.05,
  beta = 0.05,
  centered = FALSE
)
}
\arguments{
\item{psi}{A function that takes a numeric matrix \code{X} as input and returns a numeric vector.}

\item{X}{A numeric matrix of size n x d (input data).}

\item{delta_Mu}{A function of \code{X} that determines the difference between primary counterfactual outcomes.}

\item{delta_Nu}{A function of \code{X} that determines the difference between secondary counterfactual outcomes.}

\item{lambda}{A numeric scalar controlling the weight of the constraint function in the objective.}

\item{alpha}{A numeric scalar (0.05 by default) representing the constraint tolerance.}

\item{beta}{A numeric scalar (0.05 by default) controlling the sharpness of the probability function.}

\item{centered}{A logical (FALSE by default) indicating whether to apply centering in \code{sigma_beta}.}
}
\value{
A numeric scalar representing the objective function value.
}
\description{
Computes the objective function, which balances the risk function \code{R_p}
and the constraint function \code{S_p} using a parameter \code{lambda}.
}
