% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_probability.R
\name{learn_threshold}
\alias{learn_threshold}
\title{Learn Optimal Decision Threshold}
\usage{
learn_threshold(theta, X, A, Y, Xi, folds, alpha)
}
\arguments{
\item{theta}{A numeric matrix (k x d). Each row is from FW inner minimization, used to recover an extremal point for convex function construction.}

\item{X}{A matrix or data frame of covariates of size n x d (input data in \link{0,1}).}

\item{A}{A binary vector or matrix of length n indicating treatment assignment (0 or 1).}

\item{Y}{A numeric vector or matrix of length n representing primary outcomes (in \link{0, 1}).}

\item{Xi}{A numeric vector or matrix of adverse events outcomes.}

\item{folds}{A list of cross-validation folds, typically created with \code{SuperLearner::CVFolds}.}

\item{alpha}{A numeric scalar representing the constraint tolerance (in \link{0,1/2}, 0.1 by default).}
}
\value{
A numeric value corresponding to the optimal threshold chosen
from the candidate sequence.
}
\description{
This function estimates the optimal decision threshold for treatment assignment
using cross-fitted nuisance function estimates and targeted maximum likelihood
estimation (TMLE). The procedure evaluates candidate thresholds based on empirical
performance criteria, selecting the threshold that maximizes the policy value
under constraint satisfaction.
}
