% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_algorithm.R
\name{main_algorithm}
\alias{main_algorithm}
\title{Main algorithm}
\usage{
main_algorithm(
  X,
  A,
  Y,
  Xi,
  Lambdas = seq(1, 8, by = 1),
  alpha = 0.1,
  precision = 0.025,
  Betas = c(0.05, 0.5, 1, 2),
  centered = FALSE,
  Jfold = 3,
  V = 2L,
  SL.library = c("SL.mean", "SL.glm", "SL.ranger", "SL.grf"),
  root.path
)
}
\arguments{
\item{X}{A matrix or data frame of covariates.}

\item{A}{A binary vector (0/1) indicating treatment assignment.}

\item{Y}{A numeric vector or matrix of primary outcomes.}

\item{Xi}{A numeric vector or matrix of primary outcomes.}

\item{Lambdas}{A sequence of scalars controlling the weight of the constraint function in the objective (seq(1,8,by=1) by default).}

\item{alpha}{A numeric scalar representing the constraint tolerance (0.01 by default).}

\item{precision}{A numeric scalar that determines the convergence precision desired (0.025 by default).}

\item{Betas}{A vector of scalars controlling the sharpness of the probability function (c(0.05, 0.5, 1, 2) by default).}

\item{centered}{A logical value indicating whether to apply centering in \code{sigma_beta} (FALSE by default).}

\item{Jfold}{Number of folds for the whole procedure, needs to be set to 3L}

\item{V}{Number of folds inside the SuperLearner (2L by default)}

\item{SL.library}{Libraries for training learner (c("SL.mean","SL.glm","SL.ranger","SL.grf") by default)}

\item{root.path}{Path to the folder where all results are to be saved.}
}
\description{
This function runs the whole algorithm. It first checks the data, then computes the nuisance parameters using SuperLearner librarie.
Then proceed onto the
}
