% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_functions.R
\name{make_psi}
\alias{make_psi}
\title{Generate Psi Function}
\usage{
make_psi(Theta)
}
\arguments{
\item{Theta}{A numeric matrix of size k x d, where each row represents the k-th solution
\code{theta} obtained from the inner minimization in the FW algorithm.}
}
\value{
A function that computes \code{psi} for a given input matrix \code{x},
using a convex combination of past solutions.
}
\description{
Constructs a convex combination function \code{psi} based on the sequence of solutions
obtained from the Frank-Wolfe (FW) algorithm. Each new solution \code{theta} contributes
to \code{psi} in the form \eqn{2 \cdot \text{expit}(X \theta) - 1}.
}
