% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{model_Xi_linear}
\alias{model_Xi_linear}
\title{Linear treatment effect on Xi Component Function}
\usage{
model_Xi_linear(X)
}
\arguments{
\item{X}{A matrix of covariates of size n x d (input data in \verb{[0,1]}).}
}
\value{
A numeric vector with the transformed values based on covariates and treatment.
}
\description{
Computes a linear interaction term between covariates and treatment.
}
\examples{
X <- matrix(stats::runif(10*2), 10, 2)
model_Xi_linear(X)
}
