% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{model_Y_mix}
\alias{model_Y_mix}
\title{model_Y_mix: Mixed Treatment Effect on Y Component Function}
\usage{
model_Y_mix(X, A)
}
\arguments{
\item{X}{A matrix of covariates of size n x d (input data).}

\item{A}{A binary vector or matrix of length n indicating treatment assignment (-1 or 1).}
}
\value{
A numeric vector with the transformed values based on covariates and treatment.
}
\description{
Computes a linear interaction term between covariates and treatment.
}
\examples{
X <- matrix(runif(10*2), 10, 2)
A <- rep(1, 10)
model_Y_mix(X, A)
}
