% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_probability.R
\name{naive_lwr_upper_bound_estimators}
\alias{naive_lwr_upper_bound_estimators}
\title{Naive lower and upper bound estimators for policy value and constraint}
\usage{
naive_lwr_upper_bound_estimators(mu0, nu0, pi, X, A, Y, Xi, prop_score, alpha)
}
\arguments{
\item{mu0}{A fold-specific function predicting primary outcome (Y) given treatment (A) and covariates (X).}

\item{nu0}{A fold-specific function predicting adverse event outcome (Xi) given treatment (A) and covariates (X).}

\item{pi}{A numeric vector of binary decisions of length n.}

\item{X}{A matrix or data frame of covariates of size n x d (input data in \link{0,1}).}

\item{A}{A binary vector or matrix of length n indicating treatment assignment (0 or 1).}

\item{Y}{A numeric vector or matrix of length n representing primary outcomes (in \link{0, 1}).}

\item{Xi}{A numeric vector or matrix of adverse events outcomes.}

\item{prop_score}{A function that estimates the propensity score given treatment (A) and covariates (X).}

\item{alpha}{A numeric scalar representing the constraint tolerance (in \link{0,1/2}, 0.1 by default).}
}
\value{
A list containing a numeric vector of length 2:
\itemize{
\item \code{[1]}: Lower bound for the primary outcome effect.
\item \code{[2]}: Upper bound for the adverse event effect.
}
}
\description{
This function computes lower and upper confidence bounds for the policy value
and constraint, respectively, based on their plug-in estimators.
}
