% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_metric_comparison}
\alias{plot_metric_comparison}
\title{synthetic_data_plot:: Plot Synthetic Data Setting}
\usage{
plot_metric_comparison(
  data,
  metrics = c("policy_value", "constraint"),
  techniques = NULL,
  root.path
)
}
\arguments{
\item{data}{description}

\item{metrics}{A vector containing the metrics to be represented.}

\item{techniques}{}

\item{root.path}{Path to the folder where images are to be saved.}
}
\value{
Saves a plot to "Images/"Comparison_", techniques, ".pdf".
}
\description{
Generates and saves a two-panel plot:
one showing the sign of the treatment effect (\code{delta_Mu}) and the other
visualizing the magnitude of selection effect (\code{delta_Nu}) across covariates X.1 and X.2.
}
