% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_nuisance}
\alias{plot_nuisance}
\title{Visualize the true value versus the estimated version.}
\usage{
plot_nuisance(X, Var1, Var0, Var_learner, root.path, variable_name)
}
\arguments{
\item{X}{A matrix of covariates of size n x d (input data).}

\item{Var1}{Vector of treated variable values (length n).}

\item{Var0}{Vector of control variable values  (length n).}

\item{Var_learner}{A function predicting a variable (Y, Xi or propensity score) given treatment (A) and covariates (X).}

\item{root.path}{Path to the folder where images are to be saved.}

\item{variable_name}{A string precising the name of the variable ("Y", "Xi", or "prop_score")}
}
\value{
A message indicating that the image was saved.
}
\description{
Plots the density of the true values for treated and control and the estimated versions.
}
