% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_architecture.R
\name{process_results}
\alias{process_results}
\title{Evaluate a policy}
\usage{
process_results(
  theta,
  X,
  A,
  Y,
  Xi,
  mu0,
  nu0,
  prop_score,
  lambda,
  alpha,
  beta,
  centered
)
}
\arguments{
\item{theta}{Covariate matrix (n x p) for all units.}

\item{X}{Covariate matrix (n x p) for all units.}

\item{lambda}{Regularization parameter for the risk component.}

\item{alpha}{Constraint tolerance (typically between 0 and 1).}

\item{beta}{Regularization parameter for the risk component.}

\item{centered}{Logical; whether to center the features.}

\item{delta_Mu}{A list of fold-specific treatment contrast functions (e.g., T-learner estimates).}

\item{delta_Nu}{A list of fold-specific constraint contrast functions.}
}
\value{
A vector of optimized policy parameters (\code{theta}) trained across folds.
}
\description{
This function estimates a final policy by aggregating fold-specific contrast
functions (\code{Delta_mu} and \code{Delta_nu}) using cross-validation. The contrasts
are applied fold-wise, and the final policy parameters are estimated using
a policy learning algorithm (e.g., \code{FW}).
}
