% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_functions.R
\name{R_p}
\alias{R_p}
\title{Risk Function for Conditional Average Treatment Effect (CATE)}
\usage{
R_p(psi, X, delta_Mu)
}
\arguments{
\item{psi}{A function that takes a numeric matrix \code{X} as input and returns a numeric vector.}

\item{X}{A numeric matrix of size n x d (input data).}

\item{delta_Mu}{A function of \code{X} that determines the difference between primary counterfactual outcomes.}
}
\value{
A numeric scalar representing the risk function value.
}
\description{
Computes the risk function \eqn{R_p} for estimating the Conditional Average Treatment Effect (CATE).
The function minimizes the squared error between \code{psi(X)} and \code{delta_Y(X)}.
}
