% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_functions.R
\name{S_p}
\alias{S_p}
\title{Constraint Function}
\usage{
S_p(psi, X, beta, alpha, centered, delta_Nu)
}
\arguments{
\item{psi}{A function that takes a numeric matrix \code{X} as input and returns a numeric vector.}

\item{X}{A numeric matrix of size n x d (input data).}

\item{beta}{A numeric scalar controlling the sharpness of the probability function.}

\item{alpha}{A numeric scalar representing the constraint tolerance.}

\item{centered}{A logical value indicating whether to apply centering in \code{sigma_beta}.}

\item{delta_Nu}{A function of \code{X} that determines the difference between secondary counterfactual outcomes.}
}
\value{
A numeric scalar representing the constraint function value.
}
\description{
Computes the constraint function \eqn{S_p}, which ensures that the learned policy satisfies a constraint.
This function enforces a limit on the expected impact of treatment via \code{delta_Z}.
}
