% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_functions.R
\name{SGD}
\alias{SGD}
\title{Stochastic Gradient Descent (SGD) algorithm}
\usage{
SGD(
  theta_current,
  psi,
  X,
  delta_Mu,
  delta_Nu,
  lambda,
  alpha = 0.1,
  beta = 0.05,
  centered = FALSE,
  batch_prop = 1/5,
  max_iter = 1000,
  tol = 0.001,
  lr = 0.01,
  verbose = FALSE
)
}
\arguments{
\item{theta_current}{A numeric matrix of size 1 x d (intialization for parameter to estimate).}

\item{psi}{A function that takes an input \code{X} and returns a numeric vector with values in the range \code{[-1, 1]}.}

\item{X}{A matrix of covariates of size n x d (input data).}

\item{delta_Mu}{A function of \code{X} that determines the contrast between primary outcomes.}

\item{delta_Nu}{A function of \code{X} that determines the contrast between adverse event outcomes.}

\item{lambda}{A non-negative numeric scalar controlling the penalty for violating the constraint.}

\item{alpha}{A numeric scalar representing the constraint tolerance (in \link{0,1/2}, 0.1 by default).}

\item{beta}{A non-negative numeric scalar controlling the sharpness of the probability function (0.05 by default).}

\item{centered}{A logical value indicating whether to apply centering in \code{sigma_beta} (FALSE by default).}

\item{batch_prop}{Proportion of data in a batch (by default 1/5).}

\item{max_iter}{Maximum number of iterations in the SGD (by default 1e3).}

\item{tol}{Tolerance parameter (by default 1e-3).}

\item{lr}{Learning rate parameter (by default 1e-2).}

\item{verbose}{A logical value indicating whether to print progress updates. Default is \code{FALSE}.}
}
\value{
A numeric matrix of size 1 x d (optimized parameters).
}
\description{
Performs stochastic gradient descent to optimize the parameters.
}
