% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sigma_beta}
\alias{sigma_beta}
\title{Link function}
\usage{
sigma_beta(t, beta = 0.05, centered = FALSE)
}
\arguments{
\item{t}{A vector of numerics (in \code{[-1,1]}).}

\item{beta}{A non-negative numeric scalar controlling the sharpness of the probability function (0.05 by default).}

\item{centered}{A logical value indicating whether to apply centering in \code{sigma_beta} (FALSE by default).}
}
\value{
A numeric vector of treatment probabilities.
}
\description{
Link function mapping \code{[-1,1]} to \code{[0,1]}, parametrized
by \code{beta} with an optional centering.
}
