% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.grf.R
\name{SL.grf}
\alias{SL.grf}
\title{SL.grf}
\usage{
SL.grf(Y, X, newX, family, obsWeights, ...)
}
\arguments{
\item{Y}{outcome variable}

\item{X}{training dataframe}

\item{newX}{test dataframe}

\item{family}{gaussian or binomial}

\item{obsWeights}{observation-level weights}

\item{...}{not used}
}
\value{
a list containing the predictions and the fitted object
}
\description{
This function trains conditional mean of a target variable for treated and control groups
using \code{SuperLearner}, applying cross-validation to compute out-of-fold estimates.
}
