% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{synthetic_data_plot}
\alias{synthetic_data_plot}
\title{Plot synthetic data setting}
\usage{
synthetic_data_plot(delta_Mu, delta_Nu, B = 100, root.path, name)
}
\arguments{
\item{delta_Mu}{A function that computes the treatment effect (mu difference) from covariates.}

\item{delta_Nu}{A function that computes the selection effect (nu difference) from covariates.}

\item{B}{Integer, number of Monte Carlo repetitions (1e4 by default).}

\item{root.path}{Path to the folder where images are to be saved.}

\item{name}{A string to add to the end of filename.}
}
\value{
Saves a plot to "Images/synthetic_setting.pdf".
}
\description{
Generates and saves a two-panel plot:
one showing the sign of the treatment effect (\code{delta_Mu}) and the other
visualizing the magnitude of selection effect (\code{delta_Nu}) across covariates X.1 and X.2.
}
