% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternated_procedure.R
\name{update_mu}
\alias{update_mu}
\title{Update mu via augmented covariate adjustment}
\usage{
update_mu(A, X, mu0, epsilon1, theta_collection, prop_score)
}
\arguments{
\item{A}{A binary vector or matrix of length n indicating treatment assignment (0 or 1).}

\item{X}{A matrix of covariates of size n x d (input data).}

\item{mu0}{A fold-specific function predicting primary outcome (Y) given treatment (A) and covariates (X).}

\item{epsilon1}{A numeric vector of GLM coefficients for each column in \code{psi_collection}.}

\item{theta_collection}{A list of the optimal \code{theta} enabling the reconstruction of optimal \code{psi} functions.}

\item{prop_score}{A function that estimates the propensity score given treatment (A) and covariates (X).}
}
\value{
A numeric vector of updated mu on the [0,1] scale.
}
\description{
Computes updated Mu predictions by adding the bias correction for all previous solutions at X
\code{psi(X)} scaled by coefficients \code{epsilon1_collection},
then transforms back via the logistic function.
}
