% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternated_procedure.R
\name{update_nu}
\alias{update_nu}
\title{Update nu via augmented covariate adjustment}
\usage{
update_nu(
  A,
  X,
  nu0,
  epsilon2,
  theta_collection,
  prop_score,
  beta = 0.05,
  centered = FALSE
)
}
\arguments{
\item{A}{A binary vector or matrix of length n indicating treatment assignment (0 or 1).}

\item{X}{A matrix of covariates of size n x d (input data in \link{0,1}).}

\item{nu0}{A fold-specific function predicting adverse event outcome (Xi) given treatment (A) and covariates (X).}

\item{epsilon2}{A numeric vector of GLM coefficients for each column in \code{sigma_psi_collection}.}

\item{theta_collection}{A list of the optimal \code{theta} enabling the reconstruction of optimal \code{sigma_beta(psi)} functions.}

\item{prop_score}{A function that estimates the propensity score given treatment (A) and covariates (X).}

\item{beta}{A non-negative numeric scalar controlling the sharpness of the probability function (0.05 by default).}

\item{centered}{A logical value indicating whether to apply centering in \code{sigma_beta} (FALSE by default).}
}
\value{
A numeric vector of updated nu on the [0,1] scale.
}
\description{
Computes updated Nu predictions by adding the bias correction for all previous solutions at X
\code{sigma_beta(psi(X),beta,centered)} scaled by coefficients \code{epsilon2_collection},
then transforms back via the logistic function.
}
