% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{V_p}
\alias{V_p}
\title{Oracular Approximation of Value Function}
\usage{
V_p(psi, beta = 0.05, centered = FALSE, alpha = 0.1, B = 10000, seed = NA)
}
\arguments{
\item{psi}{A function that takes an input \code{X} and returns a numeric vector with values in the range \code{[-1, 1]}.}

\item{beta}{A non-negative numeric scalar controlling the sharpness of the probability function (0.05 by default).}

\item{centered}{A logical value indicating whether to apply centering in \code{sigma_beta} (FALSE by default).}

\item{alpha}{A numeric scalar representing the constraint tolerance (in \link{0,1/2}, 0.1 by default).}

\item{B}{Integer, number of Monte Carlo repetitions (1e4 by default).}

\item{seed}{Integer or NA (NA by default).}
}
\value{
A numeric scalar representing the expected primary outcome under the policy.
}
\description{
Computes the expected outcome under a policy determined by the previously optimized \code{psi(X)}.
The policy assigns treatment probabilistically based on \code{sigma_beta(psi(X))},
and the expected outcome is calculated using counterfactual outcomes.
}
