% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{V_Pn}
\alias{V_Pn}
\title{Estimation of policy value}
\usage{
V_Pn(policy, y1, y0)
}
\arguments{
\item{policy}{A numeric vector of treatment probabilities associated with \code{X} (length n).}

\item{y1}{A numeric vector or matrix of length n representing primary outcomes under treatment (in \link{0, 1}).}

\item{y0}{A numeric vector or matrix of length n representing primary outcomes under no treatment (in \link{0, 1}).}

\item{beta}{A non-negative numeric scalar controlling the sharpness of the probability function (0.05 by default).}

\item{centered}{A logical value indicating whether to apply centering in \code{sigma_beta} (FALSE by default).}
}
\value{
A numeric scalar representing the expected primary outcome under the policy.
}
\description{
Computes the expected outcome under a policy determined by the previously optimized \code{psi(X)}.
The policy assigns treatment probabilistically based on \code{sigma_beta(psi(X))},
and the expected outcome is calculated using counterfactual outcomes.
}
