% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{visual_treatment_plot}
\alias{visual_treatment_plot}
\title{Visualize Treatment Assignment Probability}
\usage{
visual_treatment_plot(
  psi_X,
  lambda,
  beta,
  centered,
  Var_X_axis,
  Var_Y_axis,
  root.path
)
}
\arguments{
\item{psi_X}{A numeric vector with values in the range \code{[-1, 1]} representing the output of \code{psi} at fixed X.}

\item{lambda}{Regularization scalar for the constraint (display purposes).}

\item{beta}{A non-negative numeric scalar controlling the sharpness of the probability function (0.05 by default).}

\item{centered}{A logical value indicating whether to apply centering in \code{sigma_beta} (FALSE by default).}

\item{Var_X_axis}{Vector of covariates values for x-axis (length n).}

\item{Var_Y_axis}{Vector of covariates values for y-axis (length n).}

\item{root.path}{Path to the folder where images are to be saved.}
}
\value{
A message indicating that the image was saved.
}
\description{
Plots the smoothed treatment assignment probability over covariates Var_X_axis and Var_Y_axis
}
