% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_data}
\alias{check_data}
\title{Check Input Data for Validity}
\usage{
check_data(Y, Xi, A, X, folds)
}
\arguments{
\item{Y}{A numeric vector or matrix of length n representing primary outcomes (in \link{0, 1}).}

\item{Xi}{A numeric vector or matrix of length n indicating adverse events (0 or 1).}

\item{A}{A binary vector or matrix of length n indicating treatment assignment (0 or 1).}

\item{X}{A matrix or data frame of covariates of size n x d (input data).}

\item{folds}{A list of cross-validation folds, typically created with \code{SuperLearner::CVFolds}.}
}
\value{
A list with elements: \code{ok} (logical), and \code{diagnoses} (character vector of issues).
}
\description{
Performs quality control checks on the input data to ensure it meets expected formats and conditions.
}
