% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimate_nu}
\alias{estimate_nu}
\title{Estimate nu}
\usage{
estimate_nu(Xi, A, X, folds, SL.library, V = 2L, threshold = 0.001)
}
\arguments{
\item{Xi}{A numeric vector or matrix of adverse events outcomes.}

\item{A}{A binary vector or matrix of length n indicating treatment assignment (0 or 1).}

\item{X}{A matrix or data frame of covariates of size n x d (input data).}

\item{folds}{A list of cross-validation folds, typically created with \code{SuperLearner::CVFolds}.}

\item{SL.library}{Vector of libraries for training SuperLearner.}

\item{V}{Number of folds inside the SuperLearner (2L by default).}

\item{threshold}{A numeric scalar that sets the minimum allowed value for upper and lower bound estimations (1e-3 by default). Constrains estimation to $\link{threshold, 1 - threshold}$.}
}
\value{
A fold-specific function predicting adverse event outcome (Xi) given treatment (A) and covariates (X)
}
\description{
This function trains conditional mean of adverse event outcome models for treated and control groups
using \code{SuperLearner}, applying cross-validation to compute out-of-fold estimates.
}
\examples{
\dontrun{
set.seed(123)
X <- matrix(rnorm(100 * 5), ncol = 5)
A <- stats::rbinom(100, 1, 0.5)
Xi <- rbinom(100, n=1, p=0.5)
JFold <- 3
folds <- folds <- SuperLearner::CVFolds(n, id = NULL,Y = Y, cvControl = SuperLearner::SuperLearner.CV.control(V = JFold, shuffle = TRUE))
SL.library <- c("SL.glm", "SL.mean", "SL.ranger")
nu_functions <- estimate_nu(Xi, A, X, folds)
# Apply a function from the list to new data:
nu_functions(1,X[1:5, ],1)  # Predict treated outcomes
}
}
