% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_functions.R
\name{FW}
\alias{FW}
\alias{FW_X}
\title{Frank-Wolfe Algorithm}
\usage{
FW(
  X,
  delta_Mu,
  delta_Nu,
  lambda,
  alpha = 0.05,
  beta = 0.05,
  centered = FALSE,
  precision = 0.05,
  verbose = TRUE
)

FW_X(
  X,
  delta_Mu,
  delta_Nu,
  lambda,
  alpha = 0.05,
  beta = 0.05,
  centered = FALSE,
  precision = 0.05,
  verbose = TRUE
)
}
\arguments{
\item{X}{A numeric matrix (input data).}

\item{delta_Mu}{A function of \code{X} that determines the difference between primary counterfactual outcomes.}

\item{delta_Nu}{A function of \code{X} that determines the difference between secondary counterfactual outcomes.}

\item{lambda}{A numeric scalar controlling the weight of the constraint function in the objective.}

\item{alpha}{A numeric scalar (0.05 by default).}

\item{beta}{A numeric scalar (0.05 by default) controlling the sharpness of the probability function.}

\item{centered}{A logical (FALSE by default) indicating whether to center the policy.}

\item{precision}{A numeric scalar (0.05 by default) that determines the convergence precision desired.}

\item{verbose}{A logical value indicating whether to print progress updates. Default is \code{TRUE}.}
}
\value{
A numeric matrix containing the optimized parameter \code{theta},
where each row represents the k-th \code{theta} solution at iteration \code{k}.
}
\description{
Implements the Frank-Wolfe optimization algorithm to iteratively refine a convex
combination function \code{psi}. At each iteration, a new solution \code{theta}
is computed via stochastic gradient descent (SGD) and added to the convex combination
in the form \eqn{2 \cdot \text{expit}(X \theta) - 1}.
}
