% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_functions.R
\name{grad_Lagrangian_p}
\alias{grad_Lagrangian_p}
\alias{grad_Lagrangian_p_X}
\title{Gradient of the Objective Function}
\usage{
grad_Lagrangian_p(
  psi,
  X,
  delta_Mu,
  delta_Nu,
  lambda,
  alpha = 0.05,
  beta = 0.05,
  centered = FALSE
)

grad_Lagrangian_p_X(
  psi_X,
  delta_Mu_X,
  delta_Nu_X,
  lambda,
  alpha = 0.05,
  beta = 0.05,
  centered = FALSE
)
}
\arguments{
\item{psi}{A function that takes a numeric matrix \code{X} as input and returns a numeric vector.}

\item{X}{A numeric matrix of size n x d (input data).}

\item{delta_Mu}{A function of \code{X} that determines the difference between primary counterfactual outcomes.}

\item{delta_Nu}{A function of \code{X} that determines the difference between secondary counterfactual outcomes.}

\item{lambda}{A numeric scalar controlling the weight of the constraint function in the objective.}

\item{alpha}{A numeric scalar (0.05 by default) representing the constraint tolerance.}

\item{beta}{A numeric scalar (0.05 by default) controlling the sharpness of the probability function.}

\item{centered}{A logical (FALSE by default) indicating whether to apply centering in \code{sigma_beta}.}
}
\value{
A numeric vector representing the gradient of the objective function with respect to \code{psi(X)}.
}
\description{
Computes the gradient of the objective function with respect to \code{psi} at X.
The gradient is used in optimization algorithms like Stochastic Gradient Descent (SGD).
}
