% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{model_Xi_linear}
\alias{model_Xi_linear}
\title{model_Xi_linear: Treatment Effect on Xi Component Function}
\usage{
model_Xi_linear(X, A)
}
\arguments{
\item{X}{A matrix of covariates of size n x d (input data).}

\item{A}{A vector indicating treatment assignment (+1 or -1) for each observation.}
}
\value{
A numeric vector with the transformed values based on covariates and treatment.
}
\description{
Computes a linear interaction term between covariates and treatment.
}
\examples{
X <- matrix(runif(10*2), 10, 2)
A <- rep(1, 10)
model_Xi_linear(X, A)
}
