% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Naive_approach.R
\name{Naive_approach_algorithm}
\alias{Naive_approach_algorithm}
\title{Main algorithm}
\usage{
Naive_approach_algorithm(
  X,
  A,
  Y,
  Xi,
  folds,
  mu0_train,
  mu0_test,
  nu0_train,
  nu0_test,
  prop_score_train,
  prop_score_test,
  Lambdas = seq(1, 8, by = 1),
  alpha = 0.1,
  precision = 0.05,
  B = c(0.05, 0.1, 0.25, 0.5),
  centered = FALSE,
  root.path
)
}
\arguments{
\item{X}{A matrix of covariates of size n x d (input data).}

\item{A}{A binary vector of size n indicating treatment assignment (0 or 1).}

\item{Y}{A numeric vector or matrix of length n representing primary outcomes (in \link{0, 1}).}

\item{Xi}{A numeric vector or matrix of length n indicating adverse events (0 or 1).}

\item{Lambdas}{A sequence of non-negative numeric scalars controlling the penalty for violating the constraint (seq(1,8,by=1) by default).}

\item{alpha}{A numeric scalar representing the constraint tolerance (in \link{0,1/2}, 0.1 by default).}

\item{precision}{A numeric scalar defining the desired convergence precision (0.05 by default). The number of Frank-Wolfe iterations (K) is inversely proportional to this value, calculated as 1/precision.}

\item{B}{A vector of non-negative scalars controlling the sharpness of the treatment probability function (c(0.05, 0.1, 0.25, 0.5) by default).}

\item{centered}{A logical value indicating whether to apply centering in \code{sigma_beta} (FALSE by default).}

\item{root.path}{Path to the folder where all results are to be saved.}
}
\description{
This function runs the whole algorithm. It first checks the data, then computes the nuisance parameters using SuperLearner librarie.
Then proceed onto the
}
