% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternated_procedure.R
\name{Optimization_Estimation}
\alias{Optimization_Estimation}
\title{Iterative Optimization Procedure}
\usage{
Optimization_Estimation(
  mu0,
  nu0,
  prop_score,
  X,
  A,
  Y,
  Xi,
  lambda,
  alpha,
  precision,
  beta = 0.05,
  centered = FALSE,
  filepath_results
)
}
\arguments{
\item{mu0}{A function that returns the conditional mean of the outcome given treatment and covariates.}

\item{nu0}{A function that returns the conditional mean of the adverse event.}

\item{prop_score}{A function that estimates the propensity score given treatment and covariates.}

\item{X}{A matrix or data frame of covariates.}

\item{A}{A binary vector (0/1) indicating treatment assignment.}

\item{Y}{A numeric vector or matrix of primary outcomes.}

\item{Xi}{A numeric vector or matrix of primary outcomes.}

\item{lambda}{A numeric scalar controlling the weight of the constraint function in the objective.}

\item{alpha}{A numeric scalar representing the constraint tolerance.}

\item{precision}{A numeric scalar that determines the convergence precision desired.}

\item{beta}{A numeric scalar controlling the sharpness of the probability function.}

\item{centered}{A logical value indicating whether to apply centering in \code{sigma_beta}.}

\item{filepath_results}{Path to the folder where intermediate and final results are saved.}
}
\value{
A list containing:
\item{iter}{The number of completed iterations.}
\item{offset_mu}{Initial logit-transformed outcome predictions.}
\item{offset_nu}{Initial logit-transformed auxiliary predictions.}
\item{psi_collection}{Matrix of covariate projections across iterations.}
\item{sigma_psi_collection}{Matrix of transformed projections across iterations.}
\item{epsilon1}{GLM coefficients from the outcome model.}
\item{epsilon2}{GLM coefficients from the auxiliary model.}
\item{theta_collection}{List of parameter vectors from each iteration of the functional weight estimation.}
}
\description{
This function performs an iterative optimization routine to correct and minimize the objective function.
It iteratively finds a solution and corrects the objective function for such optimal solution, until
two consecutive solutions do not change much.
}
\details{
This function saves intermediate results to files in order to recover progress or inspect iteration-level behavior.
If the optimization converges or the maximum number of iterations is reached, the final parameter vector \code{theta_init} is saved.
}
\examples{
# (Requires user-defined functions: mu0, nu0, prop_score, FW, make_psi, sigma_beta, update_mu_XA, update_nu_XA)
# Optimization_Estimation(mu0, nu0, prop_score, df, lambda=0.1, alpha=0.1, precision=1e-4,
#                         beta=1, centered=TRUE, folder="path/to/folder", prefix="run1")

}
