% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_functions.R
\name{R_p}
\alias{R_p}
\title{Risk function for Conditional Average Treatment Effect (CATE)}
\usage{
R_p(psi, X, delta_Mu)
}
\arguments{
\item{psi}{A function that takes an input \code{X} and returns a numeric vector with values in the range \code{[-1, 1]}.}

\item{X}{A matrix of covariates of size n x d (input data).}

\item{delta_Mu}{A function of \code{X} that determines the contrast between primary outcomes.}
}
\value{
A numeric scalar representing the risk function value.
}
\description{
Computes the risk function \eqn{R_p} for estimating the Conditional Average Treatment Effect (CATE).
The function minimizes the squared error between \code{psi(X)} and \code{delta_Y(X)}.
}
