% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_functions.R
\name{S_p}
\alias{S_p}
\title{Constraint Function}
\usage{
S_p(psi, X, beta, alpha, centered, delta_Nu)
}
\arguments{
\item{psi}{A function that takes an input \code{X} and returns a numeric vector with values in the range \code{[-1, 1]}.}

\item{X}{A matrix of covariates of size n x d (input data).}

\item{beta}{A non-negative numeric scalar controlling the sharpness of the probability function (0.05 by default).}

\item{alpha}{A numeric scalar representing the constraint tolerance (in \link{0,1/2}, 0.1 by default).}

\item{centered}{A logical value indicating whether to apply centering in \code{sigma_beta} (FALSE by default).}

\item{delta_Nu}{A function of \code{X} that determines the contrast between adverse event outcomes.}
}
\value{
A numeric scalar representing the constraint function value.
}
\description{
Computes the constraint function \eqn{S_p}, which ensures that the learned policy satisfies a constraint.
This function enforces a limit on the expected impact of treatment via \code{delta_Z}.
}
