% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_functions.R
\name{SGD}
\alias{SGD}
\alias{SGD_X}
\title{Stochastic Gradient Descent (SGD)}
\usage{
SGD(
  theta_current,
  psi,
  X,
  delta_Mu,
  delta_Nu,
  lambda,
  alpha = 0.05,
  beta = 0.05,
  centered = FALSE,
  batch_prop = 1/3,
  max_iter = 1000,
  tol = 0.001,
  lr = 0.01,
  verbose = FALSE
)

SGD_X(
  theta_current,
  psi_X,
  X,
  delta_Mu_X,
  delta_Nu_X,
  lambda,
  alpha = 0.05,
  beta = 0.05,
  centered = FALSE,
  batch_prop = 1/3,
  max_iter = 1000,
  tol = 0.001,
  lr = 0.01,
  verbose
)
}
\arguments{
\item{theta_current}{A numeric matrix of size 1 x d (intialization for parameter to estimate).}

\item{psi}{A function that takes X as input.}

\item{X}{A numeric matrix of size n x d (input data).}

\item{delta_Mu}{A function of \code{X} that determines the difference between primary counterfactual outcomes.}

\item{delta_Nu}{A function of \code{X} that determines the difference between secondary counterfactual outcomes.}

\item{lambda}{A numeric scalar controlling the weight of the constraint function in the objective.}

\item{alpha}{A numeric scalar (0.05 by default).}

\item{beta}{A numeric scalar (0.05 by default) controlling the sharpness of the probability function.}

\item{centered}{A logical (FALSE by default) indicating whether to center the policy.}

\item{batch_prop}{Proportion of data in a batch (by default 1/3).}

\item{max_iter}{Maximum number of iterations in the SGD (by default 1e3).}

\item{tol}{Tolerance parameter (by default 1e-3).}

\item{lr}{Learning rate parameter (by default 1e-2).}

\item{verbose}{A logical value indicating whether to print progress.}
}
\value{
A numeric matrix of size 1 x d (optimized parameters).
}
\description{
Performs stochastic gradient descent to optimize the parameters.
}
