% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sigma_beta}
\alias{sigma_beta}
\title{Link Function}
\usage{
sigma_beta(t, beta = 0.05, centered = FALSE)
}
\arguments{
\item{t}{A vector of numerics.}

\item{beta}{A numeric scalar (0.05 by default) controlling the sharpness of the link function.}

\item{centered}{A logical (FALSE by default) indicating whether to apply centering to the link function.}
}
\value{
A numeric vector of probabilities.
}
\description{
Link function mapping \eqn{[-1,1]} to \eqn{[0,1]}, parametrized
by \code{beta} with an optional centering.
}
