% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sigma_beta_prime}
\alias{sigma_beta_prime}
\title{Derivative of Link Function}
\usage{
sigma_beta_prime(t, beta = 0.05, centered = FALSE)
}
\arguments{
\item{t}{A vector of numerics.}

\item{beta}{A numeric scalar (0.05 by default) controlling the sharpness of the link function.}

\item{centered}{A logical (FALSE by default) indicating whether to apply centering to the link function.}
}
\value{
The derivative of \code{sigma_beta} evaluated at t.
}
\description{
Computes the derivative of the link function \code{sigma_beta},
with respect to t.
}
