% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternated_procedure.R
\name{update_nu}
\alias{update_nu}
\title{Update Nu via Augmented Covariate Adjustment}
\usage{
update_nu(
  A,
  X,
  nu0,
  epsilon2,
  theta_collection,
  prop_score,
  beta = 0.05,
  centered = FALSE
)
}
\arguments{
\item{A}{A vector of binary treatment indicators (0 or 1).}

\item{X}{A matrix or data frame of covariates.}

\item{nu0}{A function predicting the adverse event given treatment and covariates.}

\item{epsilon2}{A numeric vector of GLM coefficients for each column in \code{sigma_psi_collection}.}

\item{theta_collection}{A list of the optimal \code{theta} enabling the reconstruction of optimal \code{sigma_beta(psi)} functions.}

\item{prop_score}{A function of treatment and covariates to compute the propensity score.}

\item{beta}{A numeric scalar controlling the sharpness of the probability function.}

\item{centered}{A logical value indicating whether to apply centering in \code{sigma_beta}.}
}
\value{
A numeric vector of updated mu on the [0,1] scale.
}
\description{
Computes updated Nu predictions by adding the bias correction for all previous solutions at X
\code{sigma_beta(psi(X),beta,centered)} scaled by coefficients \code{epsilon2_collection},
then transforms back via the logistic function.
}
