% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{V_Pn}
\alias{V_Pn}
\title{Oracular Approximation of Value Function}
\usage{
V_Pn(psi, X, y1, y0, beta = 0.05, centered = FALSE)
}
\arguments{
\item{psi}{A function that takes an input \code{X} and returns a numeric vector with values in the range \code{[-1, 1]}.}

\item{X}{A matrix or data frame of covariates of size n x d (input data).}

\item{y1}{A numeric vector or matrix of length n representing primary outcomes under treatment (in \link{0, 1}).}

\item{y0}{A numeric vector or matrix of length n representing primary outcomes under no treatment (in \link{0, 1}).}

\item{beta}{A non-negative numeric scalar controlling the sharpness of the probability function (0.05 by default).}

\item{centered}{A logical value indicating whether to apply centering in \code{sigma_beta} (FALSE by default).}
}
\value{
A numeric scalar representing the expected primary outcome under the policy.
}
\description{
Computes the expected outcome under a policy determined by the previously optimized \code{psi(X)}.
The policy assigns treatment probabilistically based on \code{sigma_beta(psi(X))},
and the expected outcome is calculated using counterfactual outcomes.
}
